pyproject.toml
[project]
name = "ticketing"
version = "0.1.0"
description = "Airflow plugin used to manage external DagRun tickets."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Tickets Maintainers" }
]
dependencies = [
    "apache-airflow>=3.0",
]

[project.entry-points."airflow.plugins"]
ticketing = "ticketing.plugin:TicketingPlugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ticketing"]

[tool.hatch.build]
include = ["ticketing_plugin.py"]
